Move custom errors to separate file
